+++ Classes/Scheduler.h
@@ -20,48 +20,48 @@ public:
 
     int N, K, T, R, J;
 
-    std::vector<std::vector<std::vector<std::vector<long double>>>> s0;
+    std::vector<std::vector<std::vector<std::vector<double>>>> s0;
 
-    std::vector<std::vector<std::vector<std::vector<long double>>>> d;
-    std::vector<std::vector<std::vector<std::vector<long double>>>> expD;
+    std::vector<std::vector<std::vector<std::vector<double>>>> d;
+    std::vector<std::vector<std::vector<std::vector<double>>>> expD;
 
-    std::vector<long double> TBS;
+    std::vector<double> TBS;
     std::vector<int> userId;
     std::vector<int> firstTTI;
     std::vector<int> amountTTIs;
     std::vector<std::vector<int>> framesInTTI;
     std::vector<std::vector<bool>> dataUserInTTI;
 
-    std::vector<std::vector<std::vector<std::vector<long double>>>> p;
+    std::vector<std::vector<std::vector<std::vector<double>>>> p;
     std::vector<std::vector<std::vector<std::vector<bool>>>> b;
 
-    std::vector<std::vector<std::vector<long double>>> dFactorPerR;
-    std::vector<std::vector<long double>> dFactorPerK;
+    std::vector<std::vector<std::vector<double>>> dFactorPerR;
+    std::vector<std::vector<double>> dFactorPerK;
 
-    std::vector<std::vector<std::vector<long double>>> dPrecal;
-    std::vector<std::vector<std::vector<long double>>> sPrecal;
-    std::vector<std::vector<std::vector<std::vector<long double>>>> s;
-    std::vector<std::vector<std::vector<long double>>> st;
+    std::vector<std::vector<std::vector<double>>> dPrecal;
+    std::vector<std::vector<std::vector<double>>> sPrecal;
+    std::vector<std::vector<std::vector<std::vector<double>>>> s;
+    std::vector<std::vector<std::vector<double>>> st;
 
-    std::vector<long double> G;
+    std::vector<double> G;
 
-    Scheduler() : s0(MAXK, std::vector<std::vector<std::vector<long double>>>(MAXR, std::vector<std::vector<long double>>(MAXN, std::vector<long double>(MAXT)))),
-                  d(MAXK, std::vector<std::vector<std::vector<long double>>>(MAXN, std::vector<std::vector<long double>>(MAXR, std::vector<long double>(MAXN)))),
-                  expD(MAXK, std::vector<std::vector<std::vector<long double>>>(MAXN, std::vector<std::vector<long double>>(MAXR, std::vector<long double>(MAXN)))),
+    Scheduler() : s0(MAXK, std::vector<std::vector<std::vector<double>>>(MAXR, std::vector<std::vector<double>>(MAXN, std::vector<double>(MAXT)))),
+                  d(MAXK, std::vector<std::vector<std::vector<double>>>(MAXN, std::vector<std::vector<double>>(MAXR, std::vector<double>(MAXN)))),
+                  expD(MAXK, std::vector<std::vector<std::vector<double>>>(MAXN, std::vector<std::vector<double>>(MAXR, std::vector<double>(MAXN)))),
                   TBS(MAXJ),
                   userId(MAXJ),
                   firstTTI(MAXJ),
                   amountTTIs(MAXJ),
                   framesInTTI(MAXT, std::vector<int>()),
                   dataUserInTTI(MAXN, std::vector<bool>(MAXT)),
-                  p(MAXK, std::vector<std::vector<std::vector<long double>>>(MAXR, std::vector<std::vector<long double>>(MAXN, std::vector<long double>(MAXT)))),
+                  p(MAXK, std::vector<std::vector<std::vector<double>>>(MAXR, std::vector<std::vector<double>>(MAXN, std::vector<double>(MAXT)))),
                   b(MAXK, std::vector<std::vector<std::vector<bool>>>(MAXR, std::vector<std::vector<bool>>(MAXN, std::vector<bool>(MAXT)))),
-                  dFactorPerR(MAXK, std::vector<std::vector<long double>>(MAXR, std::vector<long double>(MAXN))),
-                  dFactorPerK(MAXK, std::vector<long double>(MAXN)),
-                  dPrecal(MAXK, std::vector<std::vector<long double>>(MAXR, std::vector<long double>(MAXN))),
-                  sPrecal(MAXK, std::vector<std::vector<long double>>(MAXR, std::vector<long double>(MAXN))),
-                  s(MAXK, std::vector<std::vector<std::vector<long double>>>(MAXR, std::vector<std::vector<long double>>(MAXN, std::vector<long double>(MAXT)))),
-                  st(MAXK, std::vector<std::vector<long double>>(MAXN, std::vector<long double>(MAXT))),
+                  dFactorPerR(MAXK, std::vector<std::vector<double>>(MAXR, std::vector<double>(MAXN))),
+                  dFactorPerK(MAXK, std::vector<double>(MAXN)),
+                  dPrecal(MAXK, std::vector<std::vector<double>>(MAXR, std::vector<double>(MAXN))),
+                  sPrecal(MAXK, std::vector<std::vector<double>>(MAXR, std::vector<double>(MAXN))),
+                  s(MAXK, std::vector<std::vector<std::vector<double>>>(MAXR, std::vector<std::vector<double>>(MAXN, std::vector<double>(MAXT)))),
+                  st(MAXK, std::vector<std::vector<double>>(MAXN, std::vector<double>(MAXT))),
                   G(MAXJ) {}
 
     void reader();
@@ -74,7 +74,7 @@ public:
 
     void computeGfromPinT(int t);
 
-    long double computeGforFrameWithoutInterferences(int j);
+    double computeGforFrameWithoutInterferences(int j);
 
     void summary();
 };
