+++ Classes/Utils.cpp
@@ -1,25 +1,25 @@
 #include "Utils.h"
 
-long double Utils::geometricMean(const std::vector<long double> &vect)
+double Utils::geometricMean(const std::vector<double> &vect)
 {
-    long double ans = 1;
+    double ans = 1;
 
     for (auto x : vect)
     {
         ans *= x;
     }
 
-    return pow(ans, (long double)1 / (long double)vect.size());
+    return pow(ans, (double)1 / (double)vect.size());
 }
 
-long double Utils::mean(const std::vector<long double> &vect)
+double Utils::mean(const std::vector<double> &vect)
 {
-    long double ans = 0;
+    double ans = 0;
 
     for (auto x : vect)
     {
         ans += x;
     }
 
-    return ans / (long double)vect.size();
+    return ans / (double)vect.size();
 }
